# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: VR GARLANDS API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /categories/{id}/products:
    get:
      operationId: getCategoryProducts
      tags: [categories]
      summary: Get products in category
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Products in category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductListResponse"
  /products:
    get:
      operationId: getProducts
      tags: [products]
      summary: List products with optional search
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: categoryId
          in: query
          schema:
            type: string
        - name: featured
          in: query
          schema:
            type: boolean
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Product list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductListResponse"
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get single product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Product detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductDetail"
        "404":
          description: Not found
...
  /orders:
    get:
      operationId: getOrders
      tags: [orders]
      summary: List user orders
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: Order list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Create a new order
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderRequest"
      responses:
        "201":
          description: Order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Product:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
          nullable: true
        price:
          type: number
        discountedPrice:
          type: number
          nullable: true
        unit:
          type: string
        image:
          type: string
        categoryId:
          type: string
        categoryName:
          type: string
        inStock:
          type: boolean
        isFeatured:
          type: boolean
        rating:
          type: number
        reviewCount:
          type: integer
        deliveryTime:
          type: integer
          description: Delivery time in minutes
        discountPercent:
          type: integer
          nullable: true
      required:
        - id
        - name
        - price
        - unit
        - image
        - categoryId
        - categoryName
        - inStock
        - isFeatured
        - rating
        - reviewCount
        - deliveryTime
    ProductDetail:
      allOf:
        - $ref: "#/components/schemas/Product"
        - type: object
          properties:
            images:
              type: array
              items:
                type: string
            reviews:
              type: array
              items:
                $ref: "#/components/schemas/Review"
            relatedProducts:
              type: array
              items:
                $ref: "#/components/schemas/Product"
    ProductListResponse:
      type: object
      properties:
        products:
          type: array
          items:
            $ref: "#/components/schemas/Product"
        total:
          type: integer
        page:
          type: integer
        totalPages:
          type: integer
      required:
        - products
        - total
        - page
        - totalPages
    Review:
      type: object
      properties:
        id:
          type: string
        userName:
          type: string
        userAvatar:
          type: string
          nullable: true
        rating:
          type: number
        comment:
          type: string
          nullable: true
        createdAt:
          type: string
      required:
        - id
        - userName
        - rating
        - createdAt
...
    Address:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        phone:
          type: string
        address:
          type: string
        city:
          type: string
        state:
          type: string
        pincode:
          type: string
        isDefault:
          type: boolean
      required:
        - id
        - name
        - phone
        - address
        - city
        - state
        - pincode
        - isDefault
...
    OrderItem:
      type: object
      properties:
        id:
          type: string
        productId:
          type: string
        name:
          type: string
        image:
          type: string
        quantity:
          type: integer
        price:
          type: number
      required:
        - id
        - productId
        - name
        - image
        - quantity
        - price
    Order:
      type: object
      properties:
        id:
          type: string
        status:
          type: string
          enum: [pending, confirmed, preparing, out_for_delivery, delivered, cancelled]
        paymentStatus:
          type: string
          enum: [pending, paid, failed, refunded]
        total:
          type: number
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        address:
          $ref: "#/components/schemas/Address"
        createdAt:
          type: string
      required:
        - id
        - status
        - paymentStatus
        - total
        - items
        - address
        - createdAt
    CreateOrderRequest:
      type: object
      properties:
        addressId:
          type: string
        couponCode:
          type: string
      required:
        - addressId
...