# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * VR GARLANDS API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Address } from './address';
import type { OrderItem } from './orderItem';
import type { OrderPaymentStatus } from './orderPaymentStatus';
import type { OrderStatus } from './orderStatus';

export interface Order {
  id: string;
  status: OrderStatus;
  paymentStatus: OrderPaymentStatus;
  total: number;
  items: OrderItem[];
  address: Address;
  createdAt: string;
}